% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{bismark_results}
\alias{bismark_results}
\title{Process methylation data from bismark}
\usage{
bismark_results(prefix, path = ".")
}
\arguments{
\item{prefix}{A \code{character}. The prefix for all file.}

\item{path}{A \code{character}. The path with the bam files generated by
\code{bismark}.}
}
\value{
A \code{data.table} with one row per position per read and methylation
state.
}
\description{
Read output from bismark and return methylation states for all reads/positions.
}
