% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{enst2symbol}
\alias{enst2symbol}
\title{Transcript to gene symbol}
\usage{
enst2symbol(enst = NULL, version = F, kallisto = F)
}
\arguments{
\item{enst}{A \code{character} vector or NULL. The annotation will be
subset for the trancripts specified.}

\item{version}{A \code{logical}. If set to TRUE, return transcript version.}

\item{kallisto}{A \code{logical}. If set to TRUE, return a table that
can be used as target_mapping for \code{sleuth_prep} or tx2gene for
\code{tximport}.}
}
\description{
Use biomaRt to obtain an annotation data.table that maps from ensembl
transcript to gene symbol.
}
