% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{filter_bm_extract}
\alias{filter_bm_extract}
\title{Filter bismark cpg files}
\usage{
filter_bm_extract(file, chrom = "X")
}
\value{
A \code{data.table} of methylation status with one row per read.
}
\description{
Filter bismark results to retain columns useful for further analysis and
keep chromosome of interest.
}
