% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{find_nearest_pos}
\alias{find_nearest_pos}
\title{Find closest position in a reference}
\usage{
find_nearest_pos(pos, ref)
}
\arguments{
\item{pos}{A \code{numeric}. The list of position to match.}

\item{ref}{The \code{numeric} vector of postions to look into.}
}
\description{
For a list of positions, find the closest match in a given reference.
}
