% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fit2dt}
\alias{fit2dt}
\title{Transform lm/glm output to data.table}
\usage{
fit2dt(fit)
}
\arguments{
\item{pos}{A \code{lm} object. The output of \code{lm} or \code{glm}}
}
\description{
For a list of positions, find the closest match in a given reference.
}
