% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{get_cpg}
\alias{get_cpg}
\title{Get a list of CpGs}
\usage{
get_cpg(version = "hg38", chr = "chrX")
}
\arguments{
\item{version}{A \code{character}. Must be 'hg38' for now.}

\item{chr}{A \code{character}. The chromosome to extract CpGs for}
}
\description{
Get CpGs ranges from UCSC and matching genes.
}
