% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASfuncs.R
\name{ggman}
\alias{ggman}
\title{Draw manhattan plot}
\usage{
ggman(
  chr,
  pos,
  pval,
  names = NULL,
  minpval = 1e-50,
  signif = 5e-08,
  lowp = 0.005
)
}
\arguments{
\item{chr}{Vector of chromosome names}

\item{pos}{Vector of genomic positions}

\item{pval}{Vector of P-values from association test}
}
\description{
Draw manhattan plot
}
