% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{GTEx_colors}
\alias{GTEx_colors}
\title{GTEx color scheme}
\usage{
GTEx_colors()
}
\value{
A \code{data.table} mapping tissue, abbreviated name and colour.
}
\description{
This function maps tissues with their associated colour in the GTEx project
}
\details{
The colours are extracted from supplementary figure 5 of
"Genetic effects on gene expression across human tissues",
Nature volume 550, pages 204–213 (12 October 2017).
}
\references{
Consortium G. (2017). Genetic effects on gene expression across human tissues. Nature 550 204–213. 10.1038/nature24277
}
