% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBS.R
\name{keepLatestLog}
\alias{keepLatestLog}
\title{Keep only the latest log for each job}
\usage{
keepLatestLog(logdir = ".")
}
\arguments{
\item{logdir}{A \code{character}. The directory with log files to cleanup.}
}
\value{
A \code{data.table}. Summary of the remaining logs.
}
\description{
This function finds all PBS log files (e/o) in a directory and keep only the
latest run for each job.
}
