% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{read_bm_summary}
\alias{read_bm_summary}
\title{Read bismark summary}
\usage{
read_bm_summary(summary)
}
\arguments{
\item{summary}{A \code{character} summary file name.}
}
\description{
Read and cleanup summary from bismark2summary
}
