% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{readKali}
\alias{readKali}
\title{Process kallisto outputs}
\usage{
readKali(bstrapdir)
}
\arguments{
\item{bstrapdir}{A \code{character}. The path of a kallisto bootstrap directory.}
}
\value{
A \code{data.table}
}
\description{
Process kallisto outputs
}
\details{
This only reads the tsv files to get raw counts.
}
