% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASfuncs.R
\name{topTableSNP}
\alias{topTableSNP}
\title{Subset a table for the top SNP in each locus}
\usage{
topTableSNP(table, pos, pval, group, locus_len = 1e+06)
}
\arguments{
\item{table}{A \code{data.table} of significantly associated SNPs, with the
columns listed in the next 3 arguments. Additional columns are preserved}

\item{pos}{A \code{character}. The name of the column of positions in \code{table}.}

\item{pval}{A \code{character}. The name of the column of p-values in \code{table}.}

\item{group}{A \code{character}. The name of the column used for grouping.
Typically phenotype. Locus are considered independently for each group.}

\item{locus_len}{A \code{numeric}. The size of loci. i.e: One SNP get returned
for each window of locus_len.}
}
\value{
A subset of the original \code{table} containing only the most
significant SNP at each locus.
}
\description{
Subset a table for the top SNP in each locus
}
