% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{tt2dt}
\alias{tt2dt}
\title{topTable to data.table}
\usage{
tt2dt(tt)
}
\arguments{
\item{tt}{}
}
\value{
A \code{data.table} with the gene names as the first column
}
\description{
Transform limma topTable output into a data.table object
}
