% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{enst2symbol}
\alias{enst2symbol}
\title{Transcript to gene symbol}
\usage{
enst2symbol(enst = NULL, version = F)
}
\arguments{
\item{enst}{A \code{character} vector or NULL. The annotation will be
subset for the trancripts specified.}
}
\description{
Use biomaRt to obtain an annotation data.table that maps from ensembl
transcript to gene symbol.
}
